fix: Chip 컴포넌트 수정 및 위치 변경
- atoms -> molcules 로 이동
- DeleteButton atoms 컴포넌트를 사용하여 구성함
- color, fontsize props 삭제

[#1]
